CON-4404 Limit product images to 5

+++ Tests/Shopware/Connect/LocalProductQueryTest.php
@@ -180,19 +180,29 @@ class LocalProductQueryTest extends ConnectTestHelper
             ],
         );
 
-        $productMedia = new Media();
-        $productMedia->setFile('http://myshop/media/image/2e/4f/tea_pavilion_product_image.jpg');
-        $variantMedia = new Media();
-        $variantMedia->setFile('http://myshop/media/image/2e/4f/tea_pavilion_variant_image.jpg');
+        $productMedia = [];
+        for ($i = 1; $i < 5; $i++) {
+            $media = new Media();
+            $media->setFile(sprintf('http://myshop/media/image/2e/4f/tea_pavilion_product_image%s.jpg', $i));
+            $productMedia[] = $media;
+        }
+
+        $variantMedia = [];
+        for ($i = 1; $i < 5; $i++) {
+            $media = new Media();
+            $media->setFile(sprintf('http://myshop/media/image/2e/4f/tea_pavilion_variant_image%s.jpg', $i));
+            $variantMedia[] = $media;
+        }
+
         $this->localMediaService->expects($this->once())
             ->method('getProductMedia')
             ->with($this->anything(), $this->productContext)
-            ->willReturn(array($productMedia, $variantMedia));
+            ->willReturn(array_merge($variantMedia, $productMedia));
 
         $this->localMediaService->expects($this->once())
             ->method('getVariantMediaList')
             ->with($this->anything(), $this->productContext)
-            ->willReturn(array($row['sku'] => array($variantMedia)));
+            ->willReturn(array($row['sku'] => $variantMedia));
 
         $expectedProduct = new Product($row);
         $expectedProduct->vendor['logo_url'] = 'http://myshop/media/image/2e/4f/tea_pavilion.jpg';
@@ -251,11 +261,17 @@ class LocalProductQueryTest extends ConnectTestHelper
         ];
 
 		$expectedProduct->images = array(
-            'http://myshop/media/image/2e/4f/tea_pavilion_product_image.jpg',
-            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image.jpg'
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image1.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image2.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image3.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image4.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_product_image1.jpg',
         );
         $expectedProduct->variantImages = array(
-            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image.jpg'
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image1.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image2.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image3.jpg',
+            'http://myshop/media/image/2e/4f/tea_pavilion_variant_image4.jpg',
         );
 
         $row['vendorName'] = $row['vendor']['name'];
@@ -295,7 +311,7 @@ class LocalProductQueryTest extends ConnectTestHelper
             'tax' => 19,
             'supplier' => 'Teapavilion',
             'mainDetail' => array(
-                'number' => '9898',
+                'number' => '9898' . rand(1, 99999),
             ),
             'filterGroupId' => $group->getId(),
             'propertyValues' => array(
@@ -354,6 +370,10 @@ class LocalProductQueryTest extends ConnectTestHelper
 
     public function tearDown()
     {
+        if (!$this->article) {
+            return;
+        }
+
         $articleId = $this->article->getId();
         $this->db->exec("DELETE FROM s_articles WHERE id = $articleId");
         $this->db->exec('DELETE FROM s_articles_details WHERE ordernumber LIKE "9898%"');
